Change font weight to an enum, shorted style and stretch enums
Preparation for upcoming attribute change.
The reason for shortening the enums was that the new attribute setup has all the font selection properties in one line, and everything starting with SKB_FONT* makes it really hard to spot the different properties.

- Changed weight to an enum
- Shortened SKB_FONT_STYLE* to SKB_STYLE*
- Shortened SKB_FONT_STRETCH* to SKB_STRETCH*

+++ src/skb_font_collection_internal.h
@@ -35,7 +35,7 @@ typedef struct skb_font_t {
 	uint8_t font_family;		// font family identifier.
 	uint8_t style;				// Normal, italic, oblique (skb_font_style_t)
 	float stretch;				// From 0.5 (ultra condensed) -> 1.0 (normal) -> 2.0 (ultra wide).
-	uint16_t weight;			// weight of the font (400 = regular).
+	int32_t weight;				// weight of the font (400 = regular).
 	uint32_t generation;		// Generation index used to identify stale handles.
 	skb_font_handle_t handle;	// Unique identifier of the font.
 } skb_font_t;
